luma_third_party: Token-based authentication
Backend changes to CrowdSTF that allow tokens to authenticate
users into using a phone for a time-based session. The system previously
required users to sign up and authenticate via username/password. This
change allows user to authenticate directly into the system for a
given amount of time with a URL token.

Change-Id: I5b93539a3b13f28b72e085fec6ae42ea96ed95e6